Update README.md
# Basic Info

This project will clean up the data from the cars_raw.csv file and then filter this data to the 75th percentile of the highest rated cars according to the consumers which are calculated using a Bayesian approach to account for the number of reviews. This filtered data will then be analyzed to determine the most common characteristics of these highly rated cars. Within the acm-coding-challenge.ipynb file I go into more detail on the methodology and equations used to analyze the data.

This was my first time using python and my first time working on a project related to data science. When I first saw this coding challenge, I was intimidated because I'd never done anything similar to this, however, I really wanted to challenge myself and expand my skills so I decided to try it out and learn a new language in the process. Using Kaggle took some time to understand and additionally it took around 4 hours to complete this project with all the research and documentation I had to go through in order to understand what I was doing. However, in the end, I did feel like I learned quite a lot about python and how to use it to manipulate and analyze datasets, and I'm glad I attempted the challenge in the first place.  

# Resources Used